timelion: Mark third party services as experimental
We don't have any mechanism in place right now to verify that the third
party backends that ship with Timelion are continuing to work as
expected, and in the event that any service introduces a backwards
compatibility break or becomes unavailable, we have no plan for how
users can workaround the issue.

This is essentially the same situation we found ourselves in with our
third party tile map provider, so we should at least come up with
solutions for this before removing the experimental labels.