[agg_response/point_series] always assign series labels to points
Since forever, the vislib has been compensating for a behavior/bug in the agg_response module where "series" labels were not assigned to points if there wasn't a "series" aggregation or multiple y-axis. To compensate for this, the vislib was assigning the yAxisLabel to the series, which had the same effect, but now that the legend is rendered in it's own tick, followed by the visualization, and that behavior was being executed during the vislib render, the legend has been flashing an empty row for a while (this was worsened by the fact that vislib rendering did not trigger a digest cycle).

These changes remove the workaround code from the vislib, and update the agg_response to always send the y-axis label as the series label when there is no series aggregation. Behavior with multiple-y-axis and multiple series aggs should be preserved.