Ensure boolean scripted field values are of boolean type
they come back from the server as ‘true’ currently.  Hopefully that
will be fixed in the future, so to future-proof this I am only
converting when the value is not of type boolean.

Fixes #8677

Need to convert both number and bool strings to their respective types.

Throw an error for unexpected types/values.