Update existing filters when negated via a visualization (#10778)
Negating an existing filter via the data table wasn't working. It works in the discover doc table because it uses the filter_manager's add method, which does some of its own existence checks and inverts a filter if it already exists. The filter_bar_click_handler, which handles creation of filters in visualizations, uses the filter_bar module directly. So I added some generic existence checks to the filter_bar itself and added code to invert existing filters if necessary.

This should solve the problem for any new visualizations that might create negated filters in the future.

Fixes #10769