Get rid of map item specific reducers (#14808) (#14812)
* get rid of funneling map entry updates through

This was one of the suggestions in the redux discussion but after
sitting with it, I decided I don’t like it.  I think it is too much
overhead and makes the code more difficult to read.  There are not so
many actions at both levels (touching the whole mapping vs touching a
single mapping object) that justifies this, IMO.

* small refactor of merging panel data logic.